+++ front/components/spaces/SpaceActionsList.tsx
@@ -1,9 +1,10 @@
-import { DataTable, Spinner, usePaginationFromUrl } from "@dust-tt/sparkle";
+import { DataTable, Spinner } from "@dust-tt/sparkle";
 import type { CellContext, ColumnDef } from "@tanstack/react-table";
 import * as React from "react";
 
 import { ACTION_BUTTONS_CONTAINER_ID } from "@app/components/spaces/SpacePageHeaders";
 import { useActionButtonsPortal } from "@app/hooks/useActionButtonsPortal";
+import { usePaginationFromUrl } from "@app/hooks/usePaginationFromUrl";
 import { useQueryParams } from "@app/hooks/useQueryParams";
 import {
   getMcpServerViewDescription,
