+++ front/components/spaces/SpaceSearchLayout.tsx
@@ -1,11 +1,6 @@
 import { DATA_SOURCE_MIME_TYPE } from "@dust-tt/client";
 import type { MenuItem } from "@dust-tt/sparkle";
-import {
-  cn,
-  ScrollableDataTable,
-  SearchInput,
-  useHashParam,
-} from "@dust-tt/sparkle";
+import { cn, ScrollableDataTable, SearchInput } from "@dust-tt/sparkle";
 import { useRouter } from "next/router";
 import React from "react";
 
@@ -20,6 +15,7 @@ import { SpaceSearchContext } from "@app/components/spaces/search/SpaceSearchCon
 import { SpacePageHeader } from "@app/components/spaces/SpacePageHeaders";
 import { useCursorPaginationForDataTable } from "@app/hooks/useCursorPaginationForDataTable";
 import { useDebounce } from "@app/hooks/useDebounce";
+import { useHashParam } from "@app/hooks/useHashParams";
 import { useSendNotification } from "@app/hooks/useNotification";
 import { useQueryParams } from "@app/hooks/useQueryParams";
 import type { NodeCandidate, UrlCandidate } from "@app/lib/connectors";
