+++ sparkle/src/hooks/index.ts
@@ -1,3 +1,1 @@
 export * from "./useCopyToClipboard";
-export * from "./useHashParams";
-export * from "./usePaginationFromUrl";
